refactor: 패키지 생성하여 관리하도록 변경
