+++ src/main/java/blakcjack/controller/BlackJackController.java
@@ -2,7 +2,7 @@ package blakcjack.controller;
 
 import blakcjack.view.InputView;
 import blakcjack.view.OutputView;
-import blakcjack.domain.BlackjackGame;
+import blakcjack.domain.blackjackgame.BlackjackGame;
 import blakcjack.domain.card.Deck;
 import blakcjack.domain.participant.Dealer;
 import blakcjack.domain.participant.Participant;
